dts: cm5: Fix ethernet phy address
CM5 features the same Broadcom BCM54213PE phy for ethernet as present on
pi5, but with a different mdio address (0 instead of 1 on pi5).

Fix this and use correct mdio address 0. While at it set
interrupt_parent to rp1_gpio instead of gpio (which is an alias to the
former) as this is the handle already used in this block.

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>